Move includes to their own folder and modify makefile.
