+++ src/server/include/http.h
@@ -87,4 +87,22 @@ public:
   std::string get_content_type(std::string);
 };
 
+HTTP::HTTP()
+{
+}
+
+HTTP::~HTTP()
+{
+}
+
+std::string HTTP::get_status_message(int http_status_code)
+{
+  return http_status_messages[http_status_code];
+}
+
+std::string HTTP::get_content_type(std::string extension)
+{
+  return content_types[extension];
+}
+
 #endif /* HTTP_H */
