+++ src/server/include/json.h
@@ -1,4 +1,21 @@
-#include "json.h"
+// json.h
+
+#ifndef JSON_H // include guard
+#define JSON_H
+
+#include <map>
+#include <string>
+#include <iostream>
+
+class JSON
+{
+private:
+public:
+  JSON();
+  ~JSON();
+
+  static std::string stringify(std::map<std::string, std::string>);
+};
 
 JSON::JSON()
 {
@@ -23,3 +40,5 @@ std::string JSON::stringify(std ::map<std::string, std::string> json)
 
   return result;
 }
+
+#endif /* JSON_H */
