+++ src/server/include/request.h
@@ -1,4 +1,34 @@
-#include "request.h"
+// request.h
+
+#ifndef REQUEST_H // include guard
+#define REQUEST_H
+
+#include <map>
+#include <vector>
+
+#include "utils.h"
+
+class Request
+{
+private:
+  /* data */
+public:
+  char as_string[1024];
+
+  std::string verb;
+
+  struct request_t
+  {
+    std::string raw_path;
+    std::vector<std::string> path;
+    std::map<std::string, std::string> query;
+  } parsed;
+
+  Request();
+  ~Request();
+
+  void parse_request();
+};
 
 Request::Request()
 {
@@ -38,3 +68,5 @@ void Request::parse_request()
   parsed.path = path_tk;
   parsed.query = query_map;
 }
+
+#endif /* REQUEST_H */
