+++ src/server/include/response.h
@@ -1,4 +1,39 @@
-#include "response.h"
+// response.h
+
+#ifndef RESPONSE_H // include guard
+#define RESPONSE_H
+
+#include <string>
+#include <stdio.h>
+#include <netinet/in.h>
+#include <map>
+
+#include "http.h"
+
+class Response
+{
+private:
+  HTTP http;
+
+  struct http_response_t
+  {
+    std::string protocol = "HTTP/1.1";
+    int status_code;
+    std::string status_message;
+    std::string content_type;
+    int content_length;
+    std::string connection_status;
+  } http_response;
+
+public:
+  Response();
+  ~Response();
+
+  void set_status_code(int);
+  void set_content_type(std::string);
+
+  std::string compose_response(std::string);
+};
 
 Response::Response()
 {
@@ -34,3 +69,5 @@ void Response::set_content_type(std::string content_type)
 {
   http_response.content_type = content_type;
 }
+
+#endif /* RESPONSE_H */
